CineMatch/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.0
pydantic-settings==2.4.0
openai==1.40.0
PyJWT==2.9.0
scikit-learn==1.5.0
pandas==2.2.0
surprise==0.1
